requirements.txt
flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
notion-client==2.3.0
python-dotenv==1.0.0